Added ability to restrict number of data points to process

+++ backend/server.py
@@ -30,7 +30,7 @@ def backtesting():
     period_length = request.args.get('period')
     capital = float(request.args.get('capital'))
     stop_loss = float(request.args.get('stopLoss'))
-    num_data = float(request.args.get('dataPoints'))
+    num_data = int(request.args.get('dataPoints'))
 
     result = backtest(coin_pair, period_length, capital, stop_loss, num_data)
